+++ packages/terra-framework-docs/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 * Added
   * Column actions example added to `terra-worklist-data-grid`.
+  
+* Changed
+  * Updated the `terra-date-time-picker` example for field label.
 
 ## 1.64.0 - (February 7, 2024)